Migrate storage from local SQLite to Postgres (Supabase)
Streamlit Community Cloud has no persistent local disk and no cron, so
local SQLite couldn't survive redeploys and there was no way to schedule
the daily refresh there. Switch to Supabase-hosted Postgres so state
persists independently of any single host, and add a GitHub Actions
workflow as the scheduler for that free deploy path.

db/init_db.py wraps psycopg2 in a thin sqlite3-compatible connection
(conn.execute with '?' placeholders, dict-like rows) so collectors/
scoring/dashboard call sites didn't need per-file rewrites. db/store.py
switches to ON CONFLICT ... DO NOTHING RETURNING id (Postgres has no
cursor.lastrowid) and defensively coerces numpy scalar values to native
Python types, since psycopg2 -- unlike sqlite3 -- doesn't reliably adapt
numpy.float64 (this was caught live: it silently broke valuation-metric
inserts against a real Supabase database).

render.yaml no longer needs a persistent disk now that state lives in
Supabase, so Render's free tier works too.

+++ collectors/market_data.py
@@ -28,7 +28,10 @@ def _pct_change(series, days_back):
     end_val = series.iloc[-1]
     if not start_val:
         return None
-    return (end_val / start_val - 1) * 100
+    # pandas/numpy scalars (e.g. numpy.float64) aren't reliably adapted by
+    # psycopg2 the way sqlite3 tolerated them — cast to native float so every
+    # value written downstream is a plain Python type.
+    return float((end_val / start_val - 1) * 100)
 
 
 def collect(conn, cfg):
